move Result to util
