+++ src/util.rs
@@ -1,3 +1,5 @@
+pub type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
+
 pub trait SplitEach {
     fn split_each(&self, n: usize) -> Vec<&Self>;
 }
